+++ thermion_dart/native/include/c_api/TVertexBuffer.h
@@ -18,6 +18,8 @@ extern "C"
     // Configure the builder
     EMSCRIPTEN_KEEPALIVE void VertexBufferBuilder_bufferCount(TVertexBufferBuilder* builder, uint8_t count);
     EMSCRIPTEN_KEEPALIVE void VertexBufferBuilder_vertexCount(TVertexBufferBuilder* builder, uint32_t count);
+    EMSCRIPTEN_KEEPALIVE void VertexBufferBuilder_enableBufferObjects(TVertexBufferBuilder* builder, bool enabled);
+    EMSCRIPTEN_KEEPALIVE TVertexBufferStorageMode VertexBufferBuilder_getStorageMode(TVertexBufferBuilder* builder);
     EMSCRIPTEN_KEEPALIVE void VertexBufferBuilder_attribute(
         TVertexBufferBuilder* builder,
         TVertexAttribute attribute,
@@ -38,7 +40,6 @@ extern "C"
 
     // Get vertex count
     EMSCRIPTEN_KEEPALIVE size_t VertexBuffer_getVertexCount(TVertexBuffer* buffer);
-
     // Set buffer data
     EMSCRIPTEN_KEEPALIVE void VertexBuffer_setBufferAt(
         TEngine* engine,
@@ -47,6 +48,12 @@ extern "C"
         void* data,
         size_t sizeInBytes,
         uint32_t byteOffset
+    );
+    EMSCRIPTEN_KEEPALIVE void VertexBuffer_setBufferObjectAt(
+        TEngine* engine,
+        TVertexBuffer* buffer,
+        uint8_t bufferIndex,
+        TBufferObject* bufferObject
     );
 
     // Destroy